Add macro for pawn movegen (#32)
Previously, a for loop was used to avoid repeated code. Now we generate the repeated code with a macro. That makes it a bit more readable and perhaps also a bit faster (benches are not conclusive towards being faster, but do show some performance gain in the correct direction).

NO FUNCTIONAL CHANGE
BENCH=64172073